storage-connector/requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
protobuf==4.23.2
psycopg==3.1.9
psycopg-binary==3.1.9
SQLAlchemy==2.0.15
SQLAlchemy-serializer==1.4.1
typing_extensions==4.6.2